feat: Add RefundController, Move refund related endpoints to RefundController, fix Optional<Double> parameter not being used in service, replaced it with .orElse(null)
